MWA/mwa_sprint3/requirements.txt
attrs==21.4.0
aws-cdk-lib==2.28.1
boto3==1.24.19
botocore==1.27.19
cattrs==22.1.0
constructs==10.1.42
exceptiongroup==1.0.0rc8
iniconfig==1.1.1
jmespath==1.0.1
jsii==1.61.0
packaging==21.3
pluggy==1.0.0
publication==0.0.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
